Add portal and API logging infrastructure

+++ MEC.Portal/Models/AdminDashboardViewModel.cs
@@ -19,7 +19,7 @@ namespace MEC.Portal.Models
     {
         public string EmployeeName { get; set; } = string.Empty;
         public string LeaveType { get; set; } = string.Empty;
-        public int RequestedDays { get; set; }
+        public decimal RequestedDays { get; set; }
         public int Status { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
