+++ MEC.Portal/Models/AdminLeaveRequestViewModel.cs
@@ -8,11 +8,11 @@ namespace MEC.Portal.Models
         public int EmployeeId { get; set; }
         public string EmployeeName { get; set; } = string.Empty;
         public string LeaveType { get; set; } = string.Empty;
-        public int RequestedDays { get; set; }
+        public decimal RequestedDays { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
         public string Reason { get; set; } = string.Empty;
         public int Status { get; set; }
-        public int RemainingLeaveDays { get; set; }
+        public decimal RemainingLeaveDays { get; set; }
     }
 }
