+++ MEC.Portal/Models/ProfileViewModel.cs
@@ -7,7 +7,7 @@ namespace MEC.Portal.Models
     {
         public EmployeePortal? Profile { get; set; }
         public int PendingAnnualLeaveCount { get; set; }
-        public int PendingAnnualLeaveDays { get; set; }
+        public decimal PendingAnnualLeaveDays { get; set; }
     }
 
     public class LeaveHistoryItemViewModel
@@ -16,10 +16,10 @@ namespace MEC.Portal.Models
         public string LeaveType { get; set; } = string.Empty;
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
-        public int RequestedDays { get; set; }
+        public decimal RequestedDays { get; set; }
         public string Reason { get; set; } = string.Empty;
         public int Status { get; set; }
-        public int RemainingLeaveDays { get; set; }
+        public decimal RemainingLeaveDays { get; set; }
         public DateTime? CreatedDate { get; set; }
         public string StatusLabel { get; set; } = string.Empty;
         public string StatusTone { get; set; } = "pending";
